services/indexer/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9
pgvector==0.3.2
pydantic-settings==2.3.4
boto3==1.34.151
pypdf==4.2.0
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.4.0
PyYAML==6.0.2
beautifulsoup4==4.12.3
lxml==5.2.2
tiktoken==0.7.0
FlagEmbedding==1.2.10
numpy==1.26.4
docling==2.41.0
